Request permissions loading spinner
