+++ src/Components/Common/UI/Texts/Texts.js
@@ -3,3 +3,4 @@ export const NO_PERMISSIONS = 'You need to accept the permissions to use our app
 export const WRONG_NETWORK =
   'You are on the wrong network, you need to use ... in order to use our app!'
 export const DEFAULT_ERROR = 'Something went wrong!'
+export const WAITING_AUTHORIZATION = 'Requesting authorization'
